Consistent UI in desktop app download page
The download buttons are now of the same style as all the other buttons on the page, instead of the retro gradient glossy things. The link for 32-bit windows downloads has been turned into a second button.

The period in the text "It even works when you're offline." has been changed to an exclamation mark.