Release 1.14.0 for new downloads
After macOS 10.15 the browsers decided to just cap the
user agent's version at 10.15 so we can't easily tell them
apart, so the macOS version detection was simplified, we'll
just give a link and tell people to scroll down if they use
those versions.

Also, clarify that the legacy versions are slower, less secure,
and that support will be removed at some unknown point.