Add SignPath test certificate signing
release-automation/build.js no longer uploads the builds to GitHub
This is now done as part of the release workflow instead.

Artifacts are now uploaded to both GitHub Actions artifacts and
the tag in case something goes wrong with the tags.

This should also mean I don't have to manually delete the blockmap
and ot her update metadata files.