Write our own GitHub release asset uploader
The old action doesn't realize that the release already exists
and so it makes a new tag, then I have to download the
assets and reupload them so that the release author doesn't
say github-actions.

Easier to write our own script for this. It's also safer in some ways
since now a misconfiguration won't start deleting assets or
mark as released early since that code just doesn't exist to run
in the first place.